Started TODO list over
Completed a functional TODO list with login and authorization.

+++ app/routers/post.py
@@ -17,9 +17,9 @@ async def get_posts(db: Session = Depends(get_db)):
 
 @router.post("/", status_code=201, response_model=schemas.Post)
 def create_post(post: schemas.PostBase, db: Session = Depends(get_db),
-                user_id: int = Depends(oauth2.get_current_user)):
-    print(user_id)
-    new_post = models.Post(**post.dict())
+                current_user: schemas.UserOut = Depends(oauth2.get_current_user)):
+    post_dict = post.dict()
+    new_post = models.Post(owner_id=current_user.id, **post_dict)
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -37,12 +37,14 @@ def get_post(id: int, db: Session = Depends(get_db)):
 
 @router.delete("/{id}", status_code=204)
 def delete_post(id: int, db: Session = Depends(get_db),
-                user_id: int = Depends(oauth2.get_current_user)):
-    print(user_id)
+                current_user: schemas.UserOut = Depends(oauth2.get_current_user)):
     post = db.query(models.Post).get({"id": id})
     if not post:
         raise HTTPException(status_code=404,
                             detail=f"post with id {id} does not exist")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=403,
+                            detail="Not Authorized to Delete this Post")
     db.delete(post)
     db.commit()
     return Response(status_code=204)
@@ -50,12 +52,16 @@ def delete_post(id: int, db: Session = Depends(get_db),
 
 @router.put("/{id}", response_model=schemas.Post)
 def update_post(id: int, post: schemas.PostBase, db: Session = Depends(get_db),
-                user_id: int = Depends(oauth2.get_current_user)):
+                current_user: int = Depends(oauth2.get_current_user)):
+
     post_query = db.query(models.Post)
     old_post = post_query.get({"id": id})
     if not old_post:
         raise HTTPException(status_code=404,
                             detail=f"post with id {id} does not exist")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=403,
+                            detail="Not Authorized to Delete this Post")
     post_query.update(post.dict())
     db.commit()
     db.refresh(old_post)
